requirements.txt
click
urwid
twisted
docker-py
attrdict
